+++ SunShine_HRMs/SunShine_HRMs.csproj
@@ -59,11 +59,11 @@
     <Compile Include="RoundButton.cs">
       <SubType>Component</SubType>
     </Compile>
-    <Compile Include="View\HR_ui_Main.cs">
+    <Compile Include="View\frmMain.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="View\HR_ui_Main.Designer.cs">
-      <DependentUpon>HR_ui_Main.cs</DependentUpon>
+    <Compile Include="View\frmMain.Designer.cs">
+      <DependentUpon>frmMain.cs</DependentUpon>
     </Compile>
     <Compile Include="RoundPicture.cs">
       <SubType>Component</SubType>
@@ -81,8 +81,8 @@
       <DependentUpon>Resources.resx</DependentUpon>
       <DesignTime>True</DesignTime>
     </Compile>
-    <EmbeddedResource Include="View\HR_ui_Main.resx">
-      <DependentUpon>HR_ui_Main.cs</DependentUpon>
+    <EmbeddedResource Include="View\frmMain.resx">
+      <DependentUpon>frmMain.cs</DependentUpon>
     </EmbeddedResource>
     <None Include="Properties\Settings.settings">
       <Generator>SettingsSingleFileGenerator</Generator>
